+++ dispatch/models.py
@@ -300,59 +300,3 @@ class DrivingHistory(models.Model):
     pub_date = models.DateTimeField(auto_now_add=True, verbose_name='작성시간')
     updated_at = models.DateTimeField(auto_now=True, verbose_name='수정시간')
     creator = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="driving_history_creator", db_column="creator_id", null=True)
-
-class DailyChecklist(models.Model):
-    submit_check = models.BooleanField(verbose_name="제출여부", null=False, default=False)
-    member = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="daily_checklist_member", null=True)
-    date = models.CharField(verbose_name="날짜", max_length=100, null=False, blank=False)
-    bus_id = models.ForeignKey(Vehicle, on_delete=models.SET_NULL, related_name="daily_checklist_bus_id", null=True)
-    oil_engine_condition = models.CharField(verbose_name="오일/엔진,부동액", max_length=100, null=False, blank=True)
-    oil_power_clutch_condition = models.CharField(verbose_name="오일/파워,클러치", max_length=100, null=False, blank=True)
-    coolant_washer_condition = models.CharField(verbose_name="냉각수,워셔액", max_length=100, null=False, blank=True)
-    external_body_condition = models.CharField(verbose_name="외부차체상태(파손확인)", max_length=100, null=False, blank=True)
-    lighting_device_condition = models.CharField(verbose_name="등화장치(실내/외)", max_length=100, null=False, blank=True)
-    blackbox_condition = models.CharField(verbose_name="블랙박스(작동여부확인)", max_length=100, null=False, blank=True)
-    tire_condition = models.CharField(verbose_name="타이어상태(나사,못)", max_length=100, null=False, blank=True)
-    interior_condition = models.CharField(verbose_name="실내상태(복도,선반,청소상태)", max_length=100, null=False, blank=True)
-    safety_belt_slide_condition = models.CharField(verbose_name="안전 벨트/슬라이드 상태", max_length=100, null=False, blank=True)
-    uniform_worn_condition = models.CharField(verbose_name="제복착용", max_length=100, null=False, blank=True)
-    pub_date = models.DateTimeField(auto_now_add=True, verbose_name='작성시간')
-    updated_at = models.DateTimeField(auto_now=True, verbose_name='수정시간')
-    creator = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="daily_checklist_creator", db_column="creator_id", null=True)
-
-class WeeklyChecklist(models.Model):
-    submit_check = models.BooleanField(verbose_name="제출여부", null=False, default=False)
-    member = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="weekly_checklist_member", null=True)
-    date = models.CharField(verbose_name="날짜", max_length=100, null=False, blank=False)
-    bus_id = models.ForeignKey(Vehicle, on_delete=models.SET_NULL, related_name="weekly_checklist_bus_id", null=True)
-    glass_tint_condition = models.CharField(verbose_name="유리/선팅", max_length=100, null=False, blank=True)
-    license_garage_record_condition = models.CharField(verbose_name="자격증/차고지증명서/운행기록증", max_length=100, null=False, blank=True)
-    tire_wheel_condition = models.CharField(verbose_name="타이어 휠 상태", max_length=100, null=False, blank=True)
-    vehicle_cleanliness_condition = models.CharField(verbose_name="차량청결(외부)", max_length=100, null=False, blank=True)
-    emergency_hammer_condition = models.CharField(verbose_name="비상망치(수량 및 야광스티커)", max_length=100, null=False, blank=True)
-    fire_extinguisher_condition = models.CharField(verbose_name="소화기(수량 및 충전상태)", max_length=100, null=False, blank=True)
-    blackbox_format_check_condition = models.CharField(verbose_name="블랙박스 포맷확인", max_length=100, null=False, blank=True)
-    pub_date = models.DateTimeField(auto_now_add=True, verbose_name='작성시간')
-    updated_at = models.DateTimeField(auto_now=True, verbose_name='수정시간')
-    creator = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="weekly_checklist_creator", db_column="creator_id", null=True)
-
-class EquipmentChecklist(models.Model):
-    submit_check = models.BooleanField(verbose_name="제출여부", null=False, default=False)
-    member = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="equipment_checklist_member", null=True)
-    date = models.CharField(verbose_name="날짜", max_length=100, null=False, blank=False)
-    bus_id = models.ForeignKey(Vehicle, on_delete=models.SET_NULL, related_name="equipment_checklist_bus_id", null=True)
-    tv_condition = models.CharField(verbose_name="TV", max_length=100, null=False, blank=True)
-    dvd_condition = models.CharField(verbose_name="DVD", max_length=100, null=False, blank=True)
-    karaoke_condition = models.CharField(verbose_name="노래방", max_length=100, null=False, blank=True)
-    floor_speaker_condition = models.CharField(verbose_name="바닥스피커", max_length=100, null=False, blank=True)
-    wireless_microphone_condition = models.CharField(verbose_name="무선마이크", max_length=100, null=False, blank=True)
-    gilseong_tech_condition = models.CharField(verbose_name="길성테크", max_length=100, null=False, blank=True)
-    floor_power_condition = models.CharField(verbose_name="바닥파워", max_length=100, null=False, blank=True)
-    inverter_condition = models.CharField(verbose_name="인버터", max_length=100, null=False, blank=True)
-    blackbox_condition = models.CharField(verbose_name="블랙박스", max_length=100, null=False, blank=True)
-    billboard_condition = models.CharField(verbose_name="전광판", max_length=100, null=False, blank=True)
-    blind_condition = models.CharField(verbose_name="블라인드", max_length=100, null=False, blank=True)
-    pub_date = models.DateTimeField(auto_now_add=True, verbose_name='작성시간')
-    updated_at = models.DateTimeField(auto_now=True, verbose_name='수정시간')
-    creator = models.ForeignKey(Member, on_delete=models.SET_NULL, related_name="equipment_checklist_creator", db_column="creator_id", null=True)
-
