+++ vehicle/views.py
@@ -6,8 +6,10 @@ from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
-from .serializers import RefuelingSerializer
-from .models import Vehicle
+from trp_drf.settings import DATE_FORMAT
+
+from .serializers import RefuelingSerializer, DailyChecklistSerializer, WeeklyChecklistSerializer, EquipmentChecklistSerializer
+from .models import Vehicle, DailyChecklist, WeeklyChecklist, EquipmentChecklist
 from crudmember.models import Category
 from humanresource.models import Member
 
@@ -59,4 +61,228 @@ class RefuelingView(APIView):
                 'message': serializer.errors,
             }
             return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+class DailyChecklistView(APIView):
+    def get(self, request, date):
+        try:
+            datetime.strptime(date, DATE_FORMAT)
+        except Exception as e:
+            response = {
+                'result': 'false',
+                'data': '1',
+                'message': {
+                    'error' : 'invalid date format'
+                },
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
         
+        try:
+            checklist = DailyChecklist.objects.filter(date=date).get(member=request.user)    
+        except DailyChecklist.DoesNotExist:
+            checklist = DailyChecklist(
+                date = date,
+                member = request.user,
+                creator = request.user
+            )
+            checklist.save()
+            
+        data = DailyChecklistSerializer(checklist).data
+        response = {
+            'result': 'true',
+            'data': data,
+            'message': ''
+        }
+        return Response(response, status=status.HTTP_200_OK)
+
+    def post(self, request, date):
+        data = request.data.copy()
+        data['member'] = request.user.id
+        data['creator'] = request.user.id
+        data['date'] = date
+
+        try:
+            datetime.strptime(date, DATE_FORMAT)
+        except Exception as e:
+            response = {
+                'result': 'false',
+                'data': '1',
+                'message': {
+                    'error' : 'invalid date format'
+                },
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            checklist = DailyChecklist.objects.filter(date=date).get(member=request.user)    
+            serializer = DailyChecklistSerializer(checklist, data=data)
+        except DailyChecklist.DoesNotExist:
+            serializer = DailyChecklistSerializer(data=data)
+        
+        if serializer.is_valid():
+            instance = serializer.save()
+            instance.submit_check = True
+            instance.save()
+            response = {
+                'result': 'true',
+                'data': serializer.data,
+                'message': ''
+            }
+            return Response(response, status=status.HTTP_200_OK)
+        response = {
+            'result': 'false',
+            'data': '2',
+            'message': {
+                'error': serializer.errors
+            }
+        }
+        return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+class WeeklyChecklistView(APIView):
+    def get(self, request, date):
+        try:
+            datetime.strptime(date, DATE_FORMAT)
+        except Exception as e:
+            response = {
+                'result': 'false',
+                'data': '1',
+                'message': {
+                    'error' : 'invalid date format'
+                },
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            checklist = WeeklyChecklist.objects.filter(date=date).get(member=request.user)    
+        except WeeklyChecklist.DoesNotExist:
+            checklist = WeeklyChecklist(
+                date = date,
+                member = request.user,
+                creator = request.user
+            )
+            checklist.save()
+            
+        data = WeeklyChecklistSerializer(checklist).data
+        response = {
+            'result': 'true',
+            'data': data,
+            'message': ''
+        }
+        return Response(response, status=status.HTTP_200_OK)
+
+    def post(self, request, date):
+        data = request.data.copy()
+        data['member'] = request.user.id
+        data['creator'] = request.user.id
+        data['date'] = date
+
+        try:
+            datetime.strptime(date, DATE_FORMAT)
+        except Exception as e:
+            response = {
+                'result': 'false',
+                'data': '1',
+                'message': {
+                    'error' : 'invalid date format'
+                },
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            checklist = WeeklyChecklist.objects.filter(date=date).get(member=request.user)    
+            serializer = WeeklyChecklistSerializer(checklist, data=data)
+        except WeeklyChecklist.DoesNotExist:
+            serializer = WeeklyChecklistSerializer(data=data)
+        
+        if serializer.is_valid():
+            instance = serializer.save()
+            instance.submit_check = True
+            instance.save()
+            response = {
+                'result': 'true',
+                'data': serializer.data,
+                'message': ''
+            }
+            return Response(response, status=status.HTTP_200_OK)
+        response = {
+            'result': 'false',
+            'data': '2',
+            'message': {
+                'error': serializer.errors
+            }
+        }
+        return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+class EquipmentChecklistView(APIView):
+    def get(self, request, date):
+        try:
+            datetime.strptime(date, DATE_FORMAT)
+        except Exception as e:
+            response = {
+                'result': 'false',
+                'data': '1',
+                'message': {
+                    'error' : 'invalid date format'
+                },
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            checklist = EquipmentChecklist.objects.filter(date=date).get(member=request.user)    
+        except EquipmentChecklist.DoesNotExist:
+            checklist = EquipmentChecklist(
+                date = date,
+                member = request.user,
+                creator = request.user
+            )
+            checklist.save()
+            
+        data = EquipmentChecklistSerializer(checklist).data
+        response = {
+            'result': 'true',
+            'data': data,
+            'message': ''
+        }
+        return Response(response, status=status.HTTP_200_OK)
+
+    def post(self, request, date):
+        data = request.data.copy()
+        data['member'] = request.user.id
+        data['creator'] = request.user.id
+        data['date'] = date
+
+        try:
+            datetime.strptime(date, DATE_FORMAT)
+        except Exception as e:
+            response = {
+                'result': 'false',
+                'data': '1',
+                'message': {
+                    'error' : 'invalid date format'
+                },
+            }
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+        
+        try:
+            checklist = EquipmentChecklist.objects.filter(date=date).get(member=request.user)    
+            serializer = EquipmentChecklistSerializer(checklist, data=data)
+        except EquipmentChecklist.DoesNotExist:
+            serializer = EquipmentChecklistSerializer(data=data)
+        
+        if serializer.is_valid():
+            instance = serializer.save()
+            instance.submit_check = True
+            instance.save()
+            response = {
+                'result': 'true',
+                'data': serializer.data,
+                'message': ''
+            }
+            return Response(response, status=status.HTTP_200_OK)
+        response = {
+            'result': 'false',
+            'data': '2',
+            'message': {
+                'error': serializer.errors
+            }
+        }
+        return Response(response, status=status.HTTP_400_BAD_REQUEST)
